Extracting Token: const token = req.cookies.token; retrieves the token from the cookies of the request object.
Token Check: If the token is not found, the function returns a 401 status code with a message indicating that the user is not authenticated and a success flag set to false.
Verifying Token: const decode = await jwt.verify(token, process.env.SECRET_KEY); verifies the token using the secret key stored in environment variables (process.env.SECRET_KEY).

Verification Check: If the token verification fails, the function returns a 401 status code with a message indicating that the token is invalid and a success flag set to false.
Setting User ID: req.id = decode.userId; sets the user ID from the decoded token payload into the req object, making it available for subsequent middleware functions and route handlers.

Calling Next Middleware: next(); calls the next middleware function in the stack if the token is successfully verified.