jobSlice Definition name: "job"
The name of the slice. It helps in generating action types and identifying this particular slice in the Redux DevTools.
initialState

An object representing the initial state of the slice. This state serves as the starting point when the slice is first created or when the state is reset.
allJobs: []: An array intended to hold all job listings available.
allAdminJobs: []: An array intended to hold jobs specifically created or managed by the admin.
singleJob: null: Holds the details of a single job. Initially set to null.
searchJobByText: "": A string for holding the current search text or query for job searching.
allAppliedJobs: []: An array to store jobs that the user has applied for.
searchedQuery: "": A string for holding the last search query made by the user.
reducers

An object defining the reducer functions that handle actions to update the state. Each function takes the current state and an action object and updates the state accordingly.
setAllJobs
Parameters: state (the current state) and action (an object containing payload with the new list of all jobs).
Purpose: Updates the allJobs property with the list of jobs provided in action.payload.
setSingleJob
Parameters: state and action (an object containing payload with details of a single job).
Purpose: Updates the singleJob property with the job details from action.payload.
setAllAdminJobs
Parameters: state and action (an object containing payload with the list of admin jobs).
Purpose: Updates the allAdminJobs property with the list of admin jobs from action.payload.
setSearchJobByText
Parameters: state and action (an object containing payload with the current search text).
Purpose: Updates the searchJobByText property with the text from action.payload.
setAllAppliedJobs
Parameters: state and action (an object containing payload with the list of applied jobs).
Purpose: Updates the allAppliedJobs property with the list of applied jobs from action.payload.
setSearchedQuery
Parameters: state and action (an object containing payload with the last search query).
Purpose: Updates the searchedQuery property with the query from action.payload.