Action Creators and Reducer export const { ... }
This line exports the action creators generated by createSlice. They are functions that can be used to dispatch actions to update the state.
setAllJobs
setSingleJob
setAllAdminJobs
setSearchJobByText
setAllAppliedJobs
setSearchedQuery
export default jobSlice.reducer;

This line exports the reducer function generated by createSlice. This reducer function will handle actions related to job state and update the state in the Redux store accordingly.
Summary
The jobSlice code defines a Redux slice for managing job-related state. It includes an initialState with properties for various job-related data and defines reducers to update these properties. The slice exports action creators and the reducer for use in other parts of the application, allowing for efficient state management and updates related to jobs.