companySlice Definition name: "company"
The name of the slice, which is used internally by Redux Toolkit to generate action types and identify this slice in the Redux DevTools.
initialState

An object that represents the initial state for this slice. It defines the default values for the slice's state properties:
singleCompany: null: Holds the details of a single company. Initially set to null, meaning no company details are available until they are set.
companies: []: An array to store a list of companies. Starts as an empty array.
searchCompanyByText: "": A string to hold the current search text or query for company searches. Initially set to an empty string.
reducers

An object defining the reducer functions that handle actions to update the state. Each function receives the current state and an action object and updates the state based on the action's payload.
setSingleCompany
Parameters: state (the current state) and action (an object containing payload with details of a single company).
Purpose: Updates the singleCompany property with the company details provided in action.payload.
setCompanies
Parameters: state and action (an object containing payload with a list of companies).
Purpose: Updates the companies property with the list of companies from action.payload.
setSearchCompanyByText
Parameters: state and action (an object containing payload with the current search text).
Purpose: Updates the searchCompanyByText property with the text from action.payload.