Action Creators and Reducer export const { setSingleCompany, setCompanies, setSearchCompanyByText } = companySlice.actions;
This line exports the action creators generated by createSlice. These action creators are functions that dispatch actions to update the state. They are:
setSingleCompany: Used to set the details of a single company.
setCompanies: Used to set the list of companies.
setSearchCompanyByText: Used to set the search text for company searches.
export default companySlice.reducer;

This line exports the reducer function generated by createSlice. This reducer function will handle actions related to company state and update the Redux store accordingly.
Summary
The companySlice code defines a Redux slice for managing state related to companies. It initializes the state with properties for storing details of a single company, a list of companies, and search text. It also provides reducers to update these properties and exports both the action creators and the reducer for use throughout the application. This slice helps manage and update company-related state in a predictable and organized manner.