applicationSlice Definition name: 'application'
The name of the slice. This name is used internally by Redux Toolkit to generate action types and to help identify this slice in the Redux DevTools. It helps in organizing and debugging the state management.
initialState

An object that represents the initial state for this slice. It defines the default values for the slice’s state properties:
applicants: null: This property is intended to store the list of applicants. It is initially set to null, indicating that there are no applicants loaded until they are set through an action.
reducers

An object containing reducer functions. These functions are used to handle actions and update the state based on the payload of those actions:
setAllApplicants
Parameters: state (the current state) and action (an object with a payload containing the list of applicants).
Purpose: Updates the applicants property of the state with the value from action.payload. This action is used to set or update the list of applicants in the state.