Action Creators and Reducer export const { setAllApplicants } = applicationSlice.actions;
This line exports the action creator function setAllApplicants. This function can be used to dispatch actions that will update the applicants state with a new list of applicants.
export default applicationSlice.reducer;

This line exports the reducer function generated by createSlice. The reducer function handles actions related to the application slice and updates the Redux store’s state accordingly.
Summary
The applicationSlice code defines a Redux slice for managing state related to job applications. It initializes the state with a property for storing the list of applicants, which starts as null. It provides a reducer (setAllApplicants) to update the list of applicants in the state. The slice exports both the action creator for updating applicants and the reducer for handling actions and updating the state in the Redux store. This helps in managing and updating application-related data in a structured and efficient manner.