Cleaning Up Search Query:
The useEffect hook is employed with an empty dependency array ([]), meaning it runs only once when the component mounts and cleans up when the component unmounts.
The cleanup function inside useEffect dispatches the setSearchedQuery action with an empty string to reset the search query in the Redux store. This ensures that when the user navigates away from the Browse component, any search query is cleared, potentially avoiding unintended search results when the user returns.