State Management:
Local State: input is a state variable managed with useState. It holds the current value of the search/filter input field.
Dispatch: dispatch is used to send actions to the Redux store.
Fetching Companies:

The useGetAllCompanies custom hook is called. This hook is responsible for fetching the list of companies from an API or data source and updating the Redux store. This hook is assumed to handle the side effects related to fetching data and might also handle storing the data in the Redux state.